test: split daemon root cli surfaces by domain
