+++ crates/daemon/tests/integration/mod.rs
@@ -4,7 +4,9 @@ use super::*;
 
 mod acp;
 mod architecture;
+mod ask_and_spec_cli_root;
 mod ask_cli;
+mod audit_cli_root;
 mod channel_surfaces_json;
 mod channel_surfaces_text;
 mod chat_cli;
@@ -32,7 +34,6 @@ mod onboard_cli;
 mod personalize_cli;
 mod plugins_cli;
 mod programmatic;
-mod root_cli;
 mod runtime_capability_cli;
 mod runtime_experiment_cli;
 mod runtime_restore_cli;
@@ -47,4 +48,5 @@ mod status_cli;
 mod tasks_cli;
 pub(crate) use managed_bridge_fixtures::*;
 mod trajectory_export_cli;
+mod validate_config_root;
 mod work_unit_cli;
